Combine building and storing config in one step
